extractor/pyproject.toml
[project]
name = "extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.75.0",
    "psycopg2-binary>=2.9.10",
]
